+++ iOS-git-osc/ProjectDetailsView.m
@@ -70,7 +70,7 @@ static NSString * const ProjectDetailsCellID = @"ProjectDetailsCell";
                                            failure:^(NSError *error) {
                                                [self.view hideToastActivity];
                                                if (error != nil) {
-                                                   [Tools toastNotification:[NSString stringWithFormat:@"网络异常，错误码：%d", error.code] inView:self.view];
+                                                   [Tools toastNotification:[NSString stringWithFormat:@"网络异常，错误码：%ld", (long)error.code] inView:self.view];
                                                } else {
                                                    [Tools toastNotification:@"网络错误" inView:self.view];
                                                }
@@ -301,7 +301,7 @@ static NSString * const ProjectDetailsCellID = @"ProjectDetailsCell";
     
     GLGitlabFailureBlock failure = ^(NSError *error) {
         if (error != nil) {
-            [Tools toastNotification:[NSString stringWithFormat:@"网络异常，错误码：%d", error.code] inView:self.view];
+            [Tools toastNotification:[NSString stringWithFormat:@"网络异常，错误码：%ld", (long)error.code] inView:self.view];
         } else {
             [Tools toastNotification:@"网络错误" inView:self.view];
         }
@@ -335,7 +335,7 @@ static NSString * const ProjectDetailsCellID = @"ProjectDetailsCell";
     
     GLGitlabFailureBlock failure = ^(NSError *error) {
         if (error != nil) {
-            [Tools toastNotification:[NSString stringWithFormat:@"网络异常，错误码：%d", error.code] inView:self.view];
+            [Tools toastNotification:[NSString stringWithFormat:@"网络异常，错误码：%ld", (long)error.code] inView:self.view];
         } else {
             [Tools toastNotification:@"网络错误" inView:self.view];
         }
